disk version was incorrectly set to 10, fixed to be 11
the wiki describing the format incorrectly stated the latest
version was 10.  the code this port was based on was actually
version 11.  currently we do not check this upon reading files
so the change is safe.  newly written files will get the correct
values.  at some point in the future we need to check the disk
version matches what we expect, logged as issue #13